Fix deployment of StylusDeployer
The deployment failed because the initialization script failed to parse
the output of forge after an update.

We also decided to simplify the script by storing the contract bytecode
instead of compiling it on the fly. The script already does that for the
cache contract, so it is reasonable to also do it for the
stylus-deployer contract. This simplifies the devnode by removing the
nitro-contract dependency.

Close STY-250
Close #7